requirements.txt
Django==2.1.7
django-cors-headers==3.0.2
djangorestframework==3.9.2
Pillow==7.1.2
pytz==2019.3
